front: split stdcm simulation report sheet
