+++ front/public/locales/en/stdcm-simulation-report-sheet.json
@@ -28,7 +28,7 @@
   "speedLimitByTag": "speed limit by tag",
   "startStop": "start",
   "stdcm": "LMR",
-  "stdcmCreation": "short term path creation",
+  "stdcmSimulation": "short term path creation",
   "stopType": "motif",
   "towedMaterial": "towed material",
   "track": "track",
