+++ front/public/locales/fr/stdcm-simulation-report-sheet.json
@@ -28,7 +28,7 @@
   "speedLimitByTag": "code de composition",
   "startStop": "départ",
   "stdcm": "LMR",
-  "stdcmCreation": "Création de sillon de dernière minute",
+  "stdcmSimulation": "Création de sillon de dernière minute",
   "stopType": "motif",
   "towedMaterial": "matériel remorqué",
   "track": "voie",
