+++ front/src/applications/stdcm/components/SimulationReportSheet/index.ts
@@ -1,0 +1,3 @@
+import SimulationReportSheet from './SimulationReportSheet';
+
+export default SimulationReportSheet;
